Move ControlDisplay to baaahs.show.live.
