+++ src/jsMain/kotlin/baaahs/app/ui/ControlsPalette.kt
@@ -1,8 +1,8 @@
 package baaahs.app.ui
 
-import baaahs.app.ui.controls.ControlDisplay
 import baaahs.app.ui.controls.SpecialControlProps
 import baaahs.app.ui.controls.control
+import baaahs.show.live.ControlDisplay
 import baaahs.show.live.OpenShow
 import baaahs.show.mutable.ShowBuilder
 import baaahs.ui.*
